sp15/06_ps_go_web_apps/7-creating-web-applications-go-m7-exercise-files/demos/after/src/converters/product.go
package converters

import (
	"models"
	"viewmodels"
)

func ConvertProductToViewModel(product models.Product) viewmodels.Product {
	result := viewmodels.Product{
		Name: product.Name(),
		DescriptionShort: product.DescriptionShort(),
		DescriptionLong: product.DescriptionLong(),
		PricePerLiter: product.PricePerLiter(),
		PricePer10Liter: product.PricePer10Liter(),
		Origin: product.Origin(),
		IsOrganic: product.IsOrganic(),
		ImageUrl: product.ImageUrl(),
		Id: product.Id(),
	}
	
	return result
}